Now ParserAPI class is in the right file
